+++ src/js/favorites.js
@@ -1,6 +1,3 @@
 import './quote-of-the-day';
 import './favorites-gallery';
-import './gallery-events';
-import { galaryState, pageFavorites } from './gallery-events';
-
-galaryState.init(pageFavorites);
+import './fav-gallery-events';
