+++ src/js/index.js
@@ -6,6 +6,3 @@ import './services/subscribe-action';
 import './gallery-events';
 import './scroll-top';
 import './exercise-popup';
-import { galaryState, pageExcercises } from './gallery-events';
-
-galaryState.init(pageExcercises);
